fixed bug is avx double using float type

+++ src/kernels/dot_impls/x86/dot_avx.cpp
@@ -13,8 +13,7 @@ double sparse_dot_double_avx(const double *vals, const uint32_t *cols,
     spira::kernel::simd::reg256_double v =
         spira::kernel::simd::load_double_256(vals + i);
 
-    float vx0 = x[cols[i]], vx1 = x[cols[i + 1]], vx2 = x[cols[i + 2]],
-          vx3 = x[cols[i + 3]];
+    double vx0 = x[cols[i]], vx1 = x[cols[i + 1]], vx2 = x[cols[i + 2]], vx3 = x[cols[i + 3]];
 
     spira::kernel::simd::reg256_double xv = _mm256_setr_pd(vx0, vx1, vx2, vx3);
 
@@ -43,15 +42,14 @@ float sparse_dot_float_avx(const float *vals, const uint32_t *cols,
           vx3 = x[cols[i + 3]], vx4 = x[cols[i + 4]], vx5 = x[cols[i + 5]],
           vx6 = x[cols[i + 6]], vx7 = x[cols[i + 7]];
 
-    spira::kernel::simd::reg256_float xv =
-        _mm256_setr_ps(vx0, vx1, vx2, vx3, vx4, vx5, vx6, vx7);
+    spira::kernel::simd::reg256_float xv = _mm256_setr_ps(vx0, vx1, vx2, vx3, vx4, vx5, vx6, vx7);
 
     acc_reg = spira::kernel::simd::fma_float_256(v, xv, acc_reg);
   }
 
   float acc = spira::kernel::simd::reduce_sum_float_256(acc_reg);
 
-  for (; i < n; i++) { // i < n, not i + 8 < n
+  for (; i < n; i++) {
     acc += vals[i] * x[cols[i]];
   }
